refactor(cli): extract tasks command module
